feat(web): redesign settings page with tabbed sections, fold profile in
Replaces the 14-card single-scroll Settings page with a tabbed layout —
General · Hardware · Behavior · Integrations · System · Advanced — using
the same shadcn Tabs primitive as Pages and Integrations. URL state via
?section= for deep-linking. The previous Profile page is folded into the
General tab (Instance Name, Appearance, Language, Time & Date, Location,
Accessibility) and the global sidebar's Profile link is removed; /profile
redirects to /settings?section=general.

Extracted GeneralSettings into UpdateIntervals + SilenceSchedule, and
extracted the six profile cards into individual settings components. Same
hooks, same API endpoints — no backend changes. Added a one-time init
guard on the polling settings effect so a late useDeferredValue settle
can't clobber user input (fixed a latent race exposed by the split).

i18n: settings.section* keys added across all 14 locales.
Tests: vitest 912/912 pass; navigation + settings E2E specs updated for
tab-aware navigation and all pass.
